# backend/src/main/resources/openapi/customer-service.yaml
openapi: 3.0.3
info:
  title: Customer Service API
  version: 1.0.0
  description: API spec for customer-service backend
servers:
  - url: http://localhost:8082
paths:
...
  /api/v1/mobile/home:
    get:
      tags: [Mobile]
      operationId: getMobileHome
      summary: Mobile home overview
      parameters:
        - name: selectedDate
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MobileHomeResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/mobile/services:
    get:
      tags: [Mobile]
      operationId: getMobileServices
      summary: Mobile services screen
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
            enum: [all, content, ads, report]
            default: all
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MobileServicesResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/mobile/notifications:
    get:
      tags: [Mobile]
      operationId: listNotifications
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 20 }
        - name: unreadOnly
          in: query
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationListResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/mobile/notifications/unread-count:
    get:
      tags: [Mobile]
      operationId: getUnreadNotificationCount
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnreadCountResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/mobile/notifications/{id}/read:
    patch:
      tags: [Mobile]
      operationId: markNotificationRead
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Marked read
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/mobile/account:
    get:
      tags: [Mobile]
      operationId: getMobileAccount
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MobileAccountResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      tags: [Mobile]
      operationId: updateMobileAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccountRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MobileAccountResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/mobile/packages:
    get:
      tags: [Mobile]
      operationId: listAvailablePackages
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PackageCatalogItem"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/mobile/package-upgrade-requests:
    post:
      tags: [Mobile]
      operationId: requestPackageUpgrade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackageUpgradeRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackageUpgradeResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/mobile/vouchers:
    get:
      tags: [Mobile]
      operationId: listMyVouchers
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserVoucherDto"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
...
  schemas:
    ErrorResponse:
      type: object
      required:
        - code
      properties:
        code:
          type: string
          example: INVALID_CREDENTIALS
        message:
          type: string
...
    SubscriptionSummary:
      type: object
      required: [id, packageCode, tier, displayTitle, startDate, endDate, status]
      properties:
        id:
          type: string
          format: uuid
        packageCode:
          type: string
        tier:
          type: string
          enum: [BASIC, PRO]
        displayTitle:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        status:
          type: string
          enum: [ACTIVE, EXPIRED, CANCELLED]
    ProgressSummary:
      type: object
      required: [overallPercent, completedPosts, completedImages, completedVideos, quotaPosts, quotaImages, quotaVideos, status]
      properties:
        overallPercent:
          type: integer
        completedPosts:
          type: integer
        completedImages:
          type: integer
        completedVideos:
          type: integer
        quotaPosts:
          type: integer
        quotaImages:
          type: integer
        quotaVideos:
          type: integer
        status:
          type: string
          enum: [pending, progress, done]
    ServiceProgressItem:
      type: object
      required: [id, name, percent, status]
      properties:
        id:
          type: string
        name:
          type: string
        icon:
          type: string
        percent:
          type: integer
        status:
          type: string
          enum: [pending, progress, done]
    ScheduleDay:
      type: object
      properties:
        date:
          type: string
          format: date
        dayOfMonth:
          type: integer
        dowIndex:
          type: integer
          description: 0=Mon .. 6=Sun
        selected:
          type: boolean
    ScheduleTask:
      type: object
      required: [id, title]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        scheduledTime:
          type: string
    MobileHomeSchedule:
      type: object
      properties:
        month:
          type: integer
        year:
          type: integer
        weekDays:
          type: array
          items:
            $ref: "#/components/schemas/ScheduleDay"
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/ScheduleTask"
    MobileHomeResponse:
      type: object
      required: [subscription, progress, services, schedule]
      properties:
        subscription:
          $ref: "#/components/schemas/SubscriptionSummary"
        progress:
          $ref: "#/components/schemas/ProgressSummary"
        services:
          type: array
          items:
            $ref: "#/components/schemas/ServiceProgressItem"
        schedule:
          $ref: "#/components/schemas/MobileHomeSchedule"
    ImplementationItem:
      type: object
      required: [id, code, title, category, currentCount, targetCount, status, updatedOn]
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        title:
          type: string
        category:
          type: string
          enum: [content, ads, report]
        currentCount:
          type: integer
        targetCount:
          type: integer
        status:
          type: string
          enum: [approved, in_progress, waiting_feedback]
        updatedOn:
          type: string
          format: date
        deliverableId:
          type: string
          format: uuid
        reviewable:
          type: boolean
    PackageServiceInfo:
      type: object
      required: [id, name, description]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
    MobileServicesResponse:
      type: object
      required: [activeSubscription, implementationItems, packageServices]
      properties:
        activeSubscription:
          $ref: "#/components/schemas/SubscriptionSummary"
        implementationItems:
          type: array
          items:
            $ref: "#/components/schemas/ImplementationItem"
        packageServices:
          type: array
          items:
            $ref: "#/components/schemas/PackageServiceInfo"
...
    NotificationDto:
      type: object
      required: [id, type, title, createdAt, read]
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [FEEDBACK_REPLY, PROMOTION, SCHEDULE]
        title:
          type: string
        body:
          type: string
        read:
          type: boolean
        createdAt:
          type: string
          format: date-time
        referenceType:
          type: string
        referenceId:
          type: string
    NotificationListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/NotificationDto"
        totalElements:
          type: integer
        totalPages:
          type: integer
    UnreadCountResponse:
      type: object
      required: [count]
      properties:
        count:
          type: integer
    MobileAccountResponse:
      type: object
      properties:
        userName:
          type: string
        fullName:
          type: string
        phone:
          type: string
        email:
          type: string
        avatarUrl:
          type: string
        subscription:
          $ref: "#/components/schemas/SubscriptionSummary"
    UpdateAccountRequest:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        email:
          type: string
        avatarUrl:
          type: string
    PackageCatalogItem:
      type: object
      required: [code, tier, label, quotaPosts, quotaImages, quotaVideos]
      properties:
        code:
          type: string
        tier:
          type: string
          enum: [BASIC, PRO]
        label:
          type: string
        quotaPosts:
          type: integer
        quotaImages:
          type: integer
        quotaVideos:
          type: integer
    PackageUpgradeRequest:
      type: object
      required: [toPackageCode]
      properties:
        toPackageCode:
          type: string
        note:
          type: string
    PackageUpgradeResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [PENDING, APPROVED, REJECTED]
    UserVoucherDto:
      type: object
      required: [id, code, title, expiresAt]
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        title:
          type: string
        description:
          type: string
        expiresAt:
          type: string
          format: date-time
        usedAt:
          type: string
          format: date-time
...